requirements-tf-2.5.2.txt
tensorflow
pandas
tensorflow-text
tensorflow-datasets
tensorflow-hub
tf-models-official==2.4.0
